@@ -22,55 +30,58 @@
   "author": "Yanfeng Tian",
   "license": "apache-2.0",
   "dependencies": {
-    "ajv": "^6.6.2",
-    "babel-runtime": "^6.26.0",
+    "@babel/runtime-corejs2": "^7.0.0",
+    "ajv": "^7.0.3",
+    "ajv-formats": "^1.5.1",
     "boom": "^7.3.0",
-    "koa": "^2.6.2",
-    "koa-body": "^4.0.4",
+    "esm": "^3.2.25",
+    "koa": "^2.13.1",
+    "koa-body": "^4.2.0",
     "koa-combine-routers": "^4.0.2",
-    "koa-logger": "^3.2.0",
-    "koa-router": "^7.4.0",
-    "massive": "^5.7.2",
-    "pg": "^7.7.1",
-    "pg-monitor": "^1.1.0",
-    "pg-promise": "^8.5.4",
-    "serverless-http": "^1.8.0",
-    "source-map-support": "^0.5.9",
-    "superagent": "^4.0.0"
+    "koa-logger": "^3.2.1",
+    "koa-router": "^10.0.0",
+    "massive": "^6.6.2",
+    "pg": "^8.5.1",
+    "pg-monitor": "^1.4.1",
+    "pg-promise": "^10.9.1",
+    "serverless-http": "^2.6.1",
+    "source-map-support": "^0.5.19",
+    "superagent": "^6.1.0"
   },
   "devDependencies": {
-    "babel-cli": "^6.26.0",
-    "babel-core": "^6.26.3",
-    "babel-loader": "^7.1.5",
-    "babel-plugin-source-map-support": "^2.0.1",
-    "babel-plugin-transform-runtime": "^6.23.0",
-    "babel-preset-env": "^1.7.0",
-    "babel-preset-stage-3": "^6.24.1",
-    "eslint": "^5.10.0",
-    "eslint-config-prettier": "^3.3.0",
-    "eslint-config-prettier-standard": "^2.0.0",
-    "eslint-config-standard": "^12.0.0",
-    "eslint-plugin-import": "^2.14.0",
-    "eslint-plugin-jest": "^22.1.2",
-    "eslint-plugin-node": "^8.0.0",
-    "eslint-plugin-prettier": "^3.0.0",
-    "eslint-plugin-promise": "^4.0.1",
-    "eslint-plugin-standard": "^4.0.0",
-    "husky": "^1.2.0",
-    "jest": "^23.6.0",
-    "lint-staged": "^8.1.0",
-    "nodemon": "^1.18.7",
-    "onchange": "^5.2.0",
-    "prettier": "^1.15.3",
-    "semistandard": "^13.0.1",
-    "serverless-jest-plugin": "^0.2.1",
-    "serverless-offline": "^3.31.3",
-    "serverless-webpack": "^5.2.0",
-    "standard": "^12.0.1",
-    "supertest": "^3.3.0",
-    "webpack": "^4.27.1",
-    "webpack-cli": "^3.1.2",
-    "webpack-node-externals": "^1.7.2"
+    "@babel/cli": "^7.12.10",
+    "@babel/core": "^7.12.10",
+    "@babel/node": "^7.0.0",
+    "@babel/plugin-transform-runtime": "^7.0.0",
+    "@babel/preset-env": "^7.12.11",
+    "babel-jest": "^26.6.3",
+    "babel-loader": "^8.2.2",
+    "babel-plugin-source-map-support": "^2.1.3",
+    "eslint": "^7.18.0",
+    "eslint-config-prettier": "^7.2.0",
+    "eslint-config-prettier-standard": "^3.0.1",
+    "eslint-config-standard": "^16.0.2",
+    "eslint-plugin-import": "^2.22.1",
+    "eslint-plugin-jest": "^24.1.3",
+    "eslint-plugin-node": "^11.1.0",
+    "eslint-plugin-prettier": "^3.3.1",
+    "eslint-plugin-promise": "^4.2.1",
+    "eslint-plugin-standard": "^4.1.0",
+    "husky": "^4.3.8",
+    "jest": "^26.6.3",
+    "lint-staged": "^10.5.3",
+    "nodemon": "^2.0.7",
+    "onchange": "^7.1.0",
+    "prettier": "^2.2.1",
+    "semistandard": "^16.0.0",
+    "serverless-jest-plugin": "^0.3.0",
+    "serverless-offline": "^6.8.0",
+    "serverless-webpack": "^5.3.5",
+    "standard": "^16.0.3",
+    "supertest": "^6.1.1",
+    "webpack": "^5.15.0",
+    "webpack-cli": "^4.3.1",
+    "webpack-node-externals": "^2.5.2"
   },
   "prettier": {
     "singleQuote": true,
