@@ -2,18 +2,32 @@ import massive from 'massive';
 // import monitor from 'pg-monitor';
 
 class Database {
-  async getDb() {
+  init() {
+    this.getDb();
+    return this;
+  }
+
+  getDb() {
     if (this.db) {
       // console.log('return db directly');
       return this.db;
     }
 
-    console.log('db creating...');
-    this.db = await massive(process.env.database);
+    console.log('db initializing...');
+
+    const cn = {
+      connectionString: process.env.database,
+      // poolSize: 10
+    };
+
+    this.db = massive(cn, {
+      documentPkType: 'uuid',
+      uuidVersion: 'v1mc',
+    });
 
     // monitor.attach(this.db);
 
-    console.log('db created.');
+    console.log('db initialized');
     return this.db;
   }
 }
