@@ -1,8 +1,8 @@
 import combineRouters from 'koa-combine-routers';
 
-import rootRouter from './roots';
-import dogRouter from './dogs';
-import customerRouter from './customers';
+import rootRouter from './roots.js';
+import dogRouter from './dogs.js';
+import customerRouter from './customers.js';
 
 const router = combineRouters(rootRouter, dogRouter, customerRouter);
 
