@@ -1,17 +1,18 @@
 import Ajv from 'ajv';
+import addFormats from 'ajv-formats';
 
-let schema = {
+const schema = {
   type: 'object',
   properties: {
     first_name: { type: 'string', minLength: 1, maxLength: 50 },
     last_name: { type: 'string', minLength: 1, maxLength: 50 },
-    email: { type: 'string', format: 'email' }
+    email: { type: 'string', format: 'email' },
   },
-  required: ['first_name', 'last_name', 'email']
+  required: ['first_name', 'last_name', 'email'],
 };
 
-let ajv = new Ajv({ allErrors: true });
-
-let customerSchema = ajv.compile(schema);
-
-export default customerSchema;
+export function getCustomerSchema() {
+  const ajv = new Ajv({ allErrors: true });
+  addFormats(ajv);
+  return ajv.compile(schema);
+}
